+++ composeApp/src/androidMain/kotlin/org/noiseplanet/noisecapture/services/storage/AndroidFileSystemService.kt
@@ -1,14 +1,13 @@
 package org.noiseplanet.noisecapture.services.storage
 
 import android.content.Context
+import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
 import org.noiseplanet.noisecapture.AndroidFilePickerEventBus
 import org.noiseplanet.noisecapture.FilePickerEvent
-import org.noiseplanet.noisecapture.util.FeatureCollection
-import org.noiseplanet.noisecapture.util.GeoJson
 import java.io.File
 import java.io.FileInputStream
 import java.io.FileOutputStream
@@ -19,39 +18,25 @@ class AndroidFileSystemService : FileSystemService, KoinComponent {
 
     // - Properties
 
+    override val dispatcher: CoroutineDispatcher
+        get() = Dispatchers.IO
+
     private val context: Context by inject()
     private val filePickerEventBus: AndroidFilePickerEventBus by inject()
 
 
     // - Public functions
 
-    override suspend fun getFileSize(fileUri: String): Long? {
-        val absolutePath = getAbsolutePath(fileUri) ?: return null
-        val file = File(absolutePath)
-        if (file.exists()) {
-            return file.length()
-        }
-        return null
-    }
-
-    override suspend fun deleteFile(fileUri: String) {
-        val absolutePath = getAbsolutePath(fileUri) ?: return
-        val file = File(absolutePath)
-        if (file.exists()) {
-            file.delete()
-        }
-    }
-
-    override suspend fun downloadFile(fileUri: String) {
+    override suspend fun download(fileUri: String) {
         // Read file contents
         val absolutePath = getAbsolutePath(fileUri) ?: return
-        val file = File(absolutePath)
+        val file = File(absolutePath.toString())
 
         // Notify activity that a new file is ready to be downloaded through event bus
         filePickerEventBus.emitEvent(FilePickerEvent(file))
     }
 
-    override suspend fun downloadFiles(fileUris: List<String>, archiveName: String) {
+    override suspend fun download(fileUris: List<String>, archiveName: String) {
         val cacheDir = context.cacheDir
         val zipFile = File(cacheDir, "$archiveName.zip")
 
@@ -60,7 +45,7 @@ class AndroidFileSystemService : FileSystemService, KoinComponent {
             ZipOutputStream(FileOutputStream(zipFile)).use { zipOut ->
                 fileUris.forEach { uri ->
                     val absolutePath = getAbsolutePath(uri) ?: return@forEach
-                    val file = File(absolutePath)
+                    val file = File(absolutePath.toString())
 
                     if (file.exists()) {
                         FileInputStream(file).use { input ->
@@ -75,17 +60,6 @@ class AndroidFileSystemService : FileSystemService, KoinComponent {
 
         // Notify activity that a new file is ready to be downloaded through event bus
         filePickerEventBus.emitEvent(FilePickerEvent(zipFile, deleteAfterUse = true))
-    }
-
-    override suspend fun downloadGeoJson(geoJson: FeatureCollection, fileName: String) {
-        val contents = GeoJson.encodeToString(geoJson)
-        val cacheDir = context.cacheDir
-        val tempFile = File(cacheDir, fileName)
-
-        // Write GeoJson data to temporary file
-        tempFile.writeText(contents)
-        // Download created file, cleaning up after use
-        filePickerEventBus.emitEvent(FilePickerEvent(tempFile, deleteAfterUse = true))
     }
 
     override fun getRootDirectory(): String {