+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/services/measurement/DefaultMeasurementService.kt
@@ -21,6 +21,9 @@ import org.noiseplanet.noisecapture.services.storage.StorageService
 import org.noiseplanet.noisecapture.services.storage.injectStorageService
 import org.noiseplanet.noisecapture.ui.theme.NoiseLevelColorRamp
 import org.noiseplanet.noisecapture.util.dbAverage
+import org.noiseplanet.noisecapture.util.geo.FeatureCollection
+import org.noiseplanet.noisecapture.util.geo.GeoJson
+import org.noiseplanet.noisecapture.util.geo.GeoJsonBuilder
 import org.noiseplanet.noisecapture.util.injectLogger
 import org.noiseplanet.noisecapture.util.isInVuMeterRange
 import org.noiseplanet.noisecapture.util.roundTo
@@ -122,10 +125,11 @@ class DefaultMeasurementService : MeasurementService, KoinComponent {
                 accumulator + (locationSequenceStorageService.getSize(sequenceId) ?: 0L)
             }
         val audioSize = measurement.recordedAudioUrl
-            ?.let { fileSystemService.getFileSize(it) }
+            ?.let { fileSystemService.size(it) }
             ?: 0L
+        val geoJsonSize = fileSystemService.size("measurement/geojson/${uuid}.geojson") ?: 0L
 
-        return measurementSize + leqSequenceSize + locationSequenceSize + audioSize
+        return measurementSize + leqSequenceSize + locationSequenceSize + audioSize + geoJsonSize
     }
 
     override fun getMeasurementFlow(uuid: String): Flow<Measurement?> {
@@ -247,11 +251,10 @@ class DefaultMeasurementService : MeasurementService, KoinComponent {
      */
     override suspend fun calculateSummary(measurement: Measurement): Measurement {
         // Read all sequence fragments to summarize data
-        val allSequenceFragments: List<LeqSequenceFragment> = measurement.leqsSequenceIds
-            .mapNotNull { leqSequenceStorageService.get(it) }
+        val leqSequence = getLeqSequenceForMeasurement(measurement.uuid)
 
         // Get a map of all LAEq values with their associated timestamp
-        val allMeasurementLaeq: Map<Long, Double> = allSequenceFragments
+        val allMeasurementLaeq: Map<Long, Double> = leqSequence
             .fold(mapOf<Long, Double>()) { accumulator, sequence ->
                 accumulator + sequence.let { it.timestamp.zip(it.laeq) }
             }.filter { (_, laeq) ->
@@ -282,7 +285,7 @@ class DefaultMeasurementService : MeasurementService, KoinComponent {
             }.toMap()
 
         // Do the same for each frequency band to compute the average level
-        val averageLeqPerFrequencyBand: Map<Int, Double> = allSequenceFragments
+        val averageLeqPerFrequencyBand: Map<Int, Double> = leqSequence
             .map { it.leqsPerThirdOctaveBand }
             .fold(mutableMapOf<Int, List<Double>>()) { accumulator, element ->
                 element.forEach { (key, values) ->
@@ -316,6 +319,30 @@ class DefaultMeasurementService : MeasurementService, KoinComponent {
         return newValue
     }
 
+    override suspend fun getMeasurementAsGeoJson(uuid: String): FeatureCollection {
+        val fileUri = "measurement/geojson/${uuid}.geojson"
+
+        return fileSystemService.readString(fileUri)?.let {
+            // If already found in filesystem, just decode locally stored representation.
+            GeoJson.decodeFromString<FeatureCollection>(it)
+        } ?: run {
+            // Calculate GeoJson representation of measurement and store it for further access
+            val leqSequence = getLeqSequenceForMeasurement(uuid)
+            val locationSequence = getLocationSequenceForMeasurement(uuid)
+            val geojson = GeoJsonBuilder.fromMeasurement(
+                leqSequence = leqSequence,
+                locationSequence = locationSequence
+            )
+            fileSystemService.writeString(
+                fileUri = fileUri,
+                text = GeoJson.encodeToString(geojson),
+            )
+            // Then return value
+            return geojson
+        }
+
+    }
+
     override suspend fun downloadRawMeasurement(uuid: String) {
         measurementStorageService.download(uuid)
     }
@@ -323,7 +350,7 @@ class DefaultMeasurementService : MeasurementService, KoinComponent {
     override suspend fun deleteMeasurementAssociatedAudio(measurement: Measurement) {
         measurement.recordedAudioUrl?.let { fileUri ->
             // Delete audio file
-            fileSystemService.deleteFile(fileUri)
+            fileSystemService.delete(fileUri)
             // And update measurement with null url
             measurementStorageService.set(
                 uuid = measurement.uuid,
@@ -343,6 +370,8 @@ class DefaultMeasurementService : MeasurementService, KoinComponent {
         measurement.locationSequenceIds.forEach {
             locationSequenceStorageService.delete(it)
         }
+        // Delete geojson representation
+        fileSystemService.delete("measurement/geojson/${measurement.uuid}.geojson")
         // Delete measurement itself
         measurementStorageService.delete(measurement.uuid)
         // Remove it from user statistics