+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/services/storage/FileSystemService.kt
@@ -1,9 +1,78 @@
 package org.noiseplanet.noisecapture.services.storage
 
-import org.noiseplanet.noisecapture.util.FeatureCollection
+import io.ktor.utils.io.core.toByteArray
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
+import kotlinx.io.buffered
+import kotlinx.io.files.Path
+import kotlinx.io.files.SystemFileSystem
+import kotlinx.io.readByteArray
 
 
+@Suppress("TooManyFunctions")
 interface FileSystemService {
+
+    // - Properties
+
+    val dispatcher: CoroutineDispatcher
+        get() = Dispatchers.Default
+
+
+    // - Public functions
+
+    /**
+     * Reads bytes from file at given URI.
+     *
+     * @param fileUri File URI.
+     * @return File contents as [ByteArray], null if file doesn't exist.
+     */
+    suspend fun read(fileUri: String): ByteArray? = with(dispatcher) {
+        val path = getAbsolutePath(fileUri) ?: return@with null
+        if (!SystemFileSystem.exists(path)) return@with null
+
+        runCatching {
+            SystemFileSystem.source(path).buffered().readByteArray()
+        }.getOrNull()
+    }
+
+    /**
+     * Reads contents of file at given URI as UTF-8 encoded String.
+     *
+     * @param fileUri File URI.
+     * @return File contents as [String], null if file doesn't exist.
+     */
+    suspend fun readString(fileUri: String): String? = with(dispatcher) {
+        read(fileUri)?.decodeToString()
+    }
+
+    /**
+     * Write given bytes to file at given URI.
+     *
+     * @param fileUri File URI.
+     * @param bytes Bytes to write.
+     * @param append If true, appends bytes to previously existing file contents. Defaults to false.
+     */
+    suspend fun write(fileUri: String, bytes: ByteArray, append: Boolean = false) =
+        with(dispatcher) {
+            val path = getAbsolutePath(fileUri) ?: return@with
+            path.parent?.let { SystemFileSystem.createDirectories(it) }
+
+            SystemFileSystem.sink(path, append)
+                .buffered()
+                .use { sink -> sink.write(bytes) }
+        }
+
+    /**
+     * Write given string to file at given URI.
+     *
+     * @param fileUri File URI.
+     * @param text String to write.
+     * @param append If true, appends bytes to previously existing file contents. Defaults to false.
+     */
+    suspend fun writeString(fileUri: String, text: String, append: Boolean = false) =
+        with(dispatcher) {
+            write(fileUri, text.toByteArray(), append)
+        }
 
     /**
      * Gets the size of the file at the given URI, in bytes.
@@ -11,38 +80,38 @@ interface FileSystemService {
      * @param fileUri File URI.
      * @return File size in bytes, null if not found.
      */
-    suspend fun getFileSize(fileUri: String): Long?
+    suspend fun size(fileUri: String): Long? = with(dispatcher) {
+        val path = getAbsolutePath(fileUri) ?: return@with null
+
+        SystemFileSystem.metadataOrNull(path)?.size
+    }
 
     /**
      * Deletes the file at the given URI, relatively to the root directory.
      *
      * @param fileUri File URI.
      */
-    suspend fun deleteFile(fileUri: String)
+    suspend fun delete(fileUri: String) = with(dispatcher) {
+        val path = getAbsolutePath(fileUri) ?: return@with
+
+        SystemFileSystem.delete(path, mustExist = false)
+    }
 
     /**
      * Opens download dialog for the given file, based on the current platform.
      *
      * @param fileUri File URI, relative to the root directory.
      */
-    suspend fun downloadFile(fileUri: String)
-
-    /**
-     * Serialises and shows download options for the given GeoJson object.
-     *
-     * @param geoJson GeoJson object
-     * @param fileName Output file name
-     */
-    suspend fun downloadGeoJson(geoJson: FeatureCollection, fileName: String)
+    suspend fun download(fileUri: String)
 
     /**
      * Zips the files at the given URIs into an archive in cache space, then lets the user download
-     * this archive through platform dependant [downloadFile] function.
+     * this archive through platform dependant [download] function.
      *
      * @param fileUris URIs of the files to download, relative to the root directory.
      * @param archiveName Name of the output archive. Defaults to "NoiseCapture_Export".
      */
-    suspend fun downloadFiles(fileUris: List<String>, archiveName: String = "NoiseCapture_Export")
+    suspend fun download(fileUris: List<String>, archiveName: String = "NoiseCapture_Export")
 
     /**
      * Returns the URI to the root directory of application files, depending on the current platform.
@@ -56,8 +125,8 @@ interface FileSystemService {
      * @param relativePath Relative path from root directory.
      * @return Absolute path (including path to root directory).
      */
-    fun getAbsolutePath(relativePath: String): String? {
+    fun getAbsolutePath(relativePath: String): Path? {
         val rootDir = getRootDirectory() ?: return null
-        return "${rootDir.replace(Regex("/*$"), "")}/$relativePath"
+        return Path("${rootDir.replace(Regex("/*$"), "")}/$relativePath")
     }
 }